Fix int overflow on arm build and add arm cross-compile to CI
armSwVer's value overflows a 32-bit int on GOOS=linux GOARCH=arm builds;
use int64 to match the pumpX2 field's actual (long) width. Add an ARM
cross-compile step to the CI build job so this class of regression is
caught automatically.

+++ pkg/settings/defaults.go
@@ -214,7 +214,7 @@ func registerPumpInfoDefaults(manager *Manager) {
 	// values below are a real captured Tandem Mobi PumpVersionResponse from
 	// pumpX2's own test fixtures (PumpVersionResponseTest.testPumpVersionResponse_Mobi).
 	registerConstant(manager, "PumpVersionRequest", map[string]interface{}{
-		"armSwVer":    3628697757,
+		"armSwVer":    int64(3628697757),
 		"mspSwVer":    0,
 		"configABits": 0,
 		"configBBits": 0,
